refactor: category page related components switch to tsx
